Regression Simulation for Linear Models

regsim.lm simulate quantities of interest from a linear regression model

# S3 method for lm
regsim(object, x, num = 1000, ...)

Arguments

object
a regression model, usually, a result of a call to lm

x
a list of explanatory variables

num
number of iterations to run

...
additional arguments passed to class-specific functions

Examples

library(regsim)

model <- lm(mpg ~ wt + cyl, data = mtcars)
sim <- regsim(model, list(wt = seq(1, 5, 0.1), cyl = mean(mtcars$cyl)))
summary(sim)

#>    (Intercept)  wt    cyl     mean        sd     2.5%      50%    97.5%
#> 1            1 1.0 6.1875 27.13881 1.6675455 23.86455 27.17837 30.25616
#> 2            1 1.1 6.1875 26.82067 1.5974271 23.68115 26.85333 29.80456
#> 3            1 1.2 6.1875 26.50253 1.5275722 23.48834 26.53729 29.35295
#> 4            1 1.3 6.1875 26.18440 1.4580188 23.29974 26.21261 28.90194
#> 5            1 1.4 6.1875 25.86626 1.3888121 23.11456 25.90168 28.45417
#> 6            1 1.5 6.1875 25.54812 1.3200068 22.92612 25.56927 28.01176
#> 7            1 1.6 6.1875 25.22998 1.2516689 22.74366 25.25497 27.56166
#> 8            1 1.7 6.1875 24.91184 1.1838795 22.55873 24.92965 27.10203
#> 9            1 1.8 6.1875 24.59371 1.1167384 22.39177 24.61517 26.67318
#> 10           1 1.9 6.1875 24.27557 1.0503699 22.20117 24.28892 26.22218
#> 11           1 2.0 6.1875 23.95743 0.9849302 22.00724 23.96618 25.80215
#> 12           1 2.1 6.1875 23.63929 0.9206176 21.84208 23.64712 25.36991
#> 13           1 2.2 6.1875 23.32115 0.8576854 21.66185 23.32445 24.92841
#> 14           1 2.3 6.1875 23.00302 0.7964610 21.47007 23.00994 24.51832
#> 15           1 2.4 6.1875 22.68488 0.7373699 21.23228 22.69393 24.11723
#> 16           1 2.5 6.1875 22.36674 0.6809677 21.04327 22.35883 23.71000
#> 17           1 2.6 6.1875 22.04860 0.6279794 20.85009 22.04602 23.27034
#> 18           1 2.7 6.1875 21.73046 0.5793423 20.61204 21.73180 22.86443
#> 19           1 2.8 6.1875 21.41233 0.5362419 20.37190 21.42175 22.47792
#> 20           1 2.9 6.1875 21.09419 0.5001116 20.10949 21.10823 22.09995
#> 21           1 3.0 6.1875 20.77605 0.4725528 19.84112 20.78639 21.71629
#> 22           1 3.1 6.1875 20.45791 0.4551255 19.59647 20.45868 21.35831
#> 23           1 3.2 6.1875 20.13977 0.4490107 19.26464 20.13870 21.05255
#> 24           1 3.3 6.1875 19.82164 0.4546651 18.92012 19.81077 20.75071
#> 25           1 3.4 6.1875 19.50350 0.4716657 18.55756 19.49768 20.44599
#> 26           1 3.5 6.1875 19.18536 0.4988538 18.15896 19.17473 20.17192
#> 27           1 3.6 6.1875 18.86722 0.5346776 17.75207 18.85150 19.92077
#> 28           1 3.7 6.1875 18.54908 0.5775322 17.34752 18.55054 19.66677
#> 29           1 3.8 6.1875 18.23095 0.6259754 16.94396 18.22894 19.43372
#> 30           1 3.9 6.1875 17.91281 0.6788117 16.50483 17.90768 19.20115
#> 31           1 4.0 6.1875 17.59467 0.7350944 16.10836 17.59494 18.97332
#> 32           1 4.1 6.1875 17.27653 0.7940911 15.66454 17.27138 18.76297
#> 33           1 4.2 6.1875 16.95839 0.8552403 15.22160 16.95711 18.55377
#> 34           1 4.3 6.1875 16.64026 0.9181120 14.75527 16.63908 18.36227
#> 35           1 4.4 6.1875 16.32212 0.9823756 14.33193 16.33701 18.16166
#> 36           1 4.5 6.1875 16.00398 1.0477749 13.88089 16.01876 17.96275
#> 37           1 4.6 6.1875 15.68584 1.1141100 13.42985 15.70724 17.75090
#> 38           1 4.7 6.1875 15.36770 1.1812233 12.97881 15.38705 17.54411
#> 39           1 4.8 6.1875 15.04957 1.2489892 12.52913 15.06488 17.34033
#> 40           1 4.9 6.1875 14.73143 1.3173071 12.08743 14.74763 17.16459
#> 41           1 5.0 6.1875 14.41329 1.3860954 11.64132 14.43328 16.98951