Regression Simulation for Generalized Linear Models

regsim.glm simulates quantities of interest from a generalized linear model

# S3 method for glm
regsim(object, x, num = 1000, link = logit, ...)

Arguments

object
a regression model, usually, a result of a call to glm

x
a list of explanatory variables

num
number of iterations to run

link
model link function

...
additional arguments passed to class-specific functions

Examples

library(regsim)

y <- swiss$Fertility > median(swiss$Fertility)
model <- glm(y ~ Education + Agriculture, family = "binomial", data = swiss)
x <- list(
  Agriculture = seq(1, 100, 5),
  Education = mean(swiss$Education)
)
sim <- regsim(model, x)
summary(sim)

#>    Education Agriculture      mean         sd       2.5%       50%     97.5%
#> 1   10.97872           1 0.6058416 0.18914785 0.22225969 0.6236746 0.9125982
#> 2   10.97872           6 0.5910140 0.17754410 0.23243441 0.6077768 0.8916673
#> 3   10.97872          11 0.5751576 0.16513549 0.24483673 0.5855008 0.8640736
#> 4   10.97872          16 0.5582996 0.15204174 0.26303755 0.5640366 0.8321044
#> 5   10.97872          21 0.5405049 0.13850297 0.27158539 0.5446202 0.7988124
#> 6   10.97872          26 0.5218828 0.12493689 0.28274143 0.5212124 0.7678111
#> 7   10.97872          31 0.5025915 0.11201013 0.29132827 0.5027772 0.7382362
#> 8   10.97872          36 0.4828371 0.10070326 0.29972977 0.4833239 0.7019393
#> 9   10.97872          41 0.4628667 0.09229010 0.29841198 0.4607184 0.6553078
#> 10  10.97872          46 0.4429546 0.08806652 0.28600386 0.4378075 0.6295652
#> 11  10.97872          51 0.4233816 0.08875883 0.26272456 0.4191530 0.6000553
#> 12  10.97872          56 0.4044122 0.09404027 0.23874031 0.3964575 0.5974851
#> 13  10.97872          61 0.3862732 0.10273835 0.20507457 0.3791175 0.6005325
#> 14  10.97872          66 0.3691385 0.11348421 0.16965486 0.3586850 0.6072546
#> 15  10.97872          71 0.3531230 0.12516023 0.14016119 0.3416929 0.6271709
#> 16  10.97872          76 0.3382851 0.13700638 0.11491669 0.3201892 0.6500415
#> 17  10.97872          81 0.3246351 0.14856147 0.09288251 0.3010565 0.6687899
#> 18  10.97872          86 0.3121464 0.15957292 0.07595796 0.2819459 0.6890683
#> 19  10.97872          91 0.3007668 0.16992185 0.06042827 0.2678169 0.6996690
#> 20  10.97872          96 0.2904279 0.17957036 0.04930531 0.2499461 0.7161089